+++ grapesy.cabal
@@ -203,36 +203,29 @@ test-suite test-grapesy
       exitcode-stdio-1.0
   hs-source-dirs:
       test-grapesy
-      test-common
   main-is:
       Main.hs
   other-modules:
       Paths_grapesy
-      Test.Driver.Dialogue.Definition
       Test.Driver.Dialogue.TestClock
-      Test.Util.ClientServer
       Test.Util.PrettyVal
   build-depends:
       -- Internal dependencies
     , grapesy
   build-depends:
       -- External dependencies
-    , async            >= 2.2  && < 2.3
-    , bytestring       >= 0.10 && < 0.12
-    , containers       >= 0.6  && < 0.7
-    , contra-tracer    >= 0.2  && < 0.3
-    , data-default     >= 0.7  && < 0.8
-    , exceptions       >= 0.10 && < 0.11
+    , async         >= 2.2  && < 2.3
+    , containers    >= 0.6  && < 0.7
+    , contra-tracer >= 0.2  && < 0.3
+    , data-default  >= 0.7  && < 0.8
+    , exceptions    >= 0.10 && < 0.11
     , http2
-    , mtl              >= 2.2  && < 2.4
-    , pretty-show      >= 1.10 && < 1.11
-    , QuickCheck       >= 2.14 && < 2.15
-    , stm              >= 2.5  && < 2.6
-    , tasty            >= 1.4  && < 1.6
-    , tasty-hunit      >= 0.10 && < 0.11
-    , tasty-quickcheck >= 0.10 && < 0.11
-    , text             >= 1.2  && < 2.1
-    , tls              >= 1.5  && < 1.8
+    , mtl           >= 2.2  && < 2.4
+    , pretty-show   >= 1.10 && < 1.11
+    , QuickCheck    >= 2.14 && < 2.15
+    , stm           >= 2.5  && < 2.6
+    , text          >= 1.2  && < 2.1
+    , tls           >= 1.5  && < 1.8
 
 Flag build-demo
   description: Build the demo
